android example renamed and project updated
