+++ android-app-plugin/.gitignore
