+++ android-app-plugin/app/.gitignore
