+++ android-app-plugin/app/build.gradle
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    namespace 'com.shurjopay.android.android_app_kotlin_plugin'
+    namespace 'com.shurjopay.android.android_app'
     compileSdk 33
 
     defaultConfig {
-        applicationId "com.shurjopay.android.android_app_kotlin_plugin"
+        applicationId "com.shurjopay.android.android_app"
         minSdk 21
         targetSdk 33
         versionCode 1
@@ -45,5 +45,5 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     // shurjopay plugin
-    implementation 'com.github.filelucker:test-plugin:v1.2.1'
+    implementation 'com.github.filelucker:test-plugin:v1.2.2'
 }
