+++ android-app-plugin/app/src/main/AndroidManifest.xml
