+++ android-app-plugin/app/src/main/java/com/shurjopay/android/android_app_example/MainActivity.kt
@@ -1,16 +1,15 @@
-package com.shurjopay.android.android_app_kotlin_plugin
+package com.shurjopay.android.android_app_example
 
+import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
-import com.shurjopay.android.android_app_kotlin_plugin.databinding.ActivityMainBinding
+import androidx.annotation.RequiresApi
+import com.shurjopay.android.android_app.databinding.ActivityMainBinding
 import com.shurjopay.android.example.utils.Constants.Companion.SHURJOPAY_API
 import com.shurjopay.android.example.utils.Constants.Companion.SP_PASS
 import com.shurjopay.android.example.utils.Constants.Companion.SP_USER
-import sp.plugin.android.v2.model.ShurjopayConfigs
-import sp.plugin.android.v2.model.ShurjopayException
-import sp.plugin.android.v2.model.ShurjopayRequestModel
-import sp.plugin.android.v2.model.ShurjopaySuccess
+import sp.plugin.android.v2.model.*
 import sp.plugin.android.v2.payment.PaymentResultListener
 import sp.plugin.android.v2.payment.Shurjopay
 import java.util.*
@@ -43,29 +42,19 @@ class MainActivity : AppCompatActivity() {
 
     private fun pay() {
 
-        val data = ShurjopayRequestModel(
+        val data = PaymentReq(
             "sp",
-            "BDT",
             binding.amountLayout.editText?.text.toString().toDouble(),
             "NOK" + Random().nextInt(1000000),
-            null,
-            null,
+            "BDT",
             binding.nameLayout.editText?.text.toString(),
-            binding.phoneLayout.editText?.text.toString(),
-            null,
             binding.addressLayout.editText?.text.toString(),
+            binding.phoneLayout.editText?.text.toString(),
             binding.cityLayout.editText?.text.toString(),
-            null,
-            null,
-            null,
-            "https://www.sandbox.shurjopayment.com/return_url",
-            "https://www.sandbox.shurjopayment.com/cancel_url",
-            "127.0.0.1",
-            "value-of-1",
-            "value-of-2",
-            "value-of-3",
-            "value-of-4"
+            "1200",
+            "m.zaman000@gmail.com",
         )
+
 
         shurjopay?.makePayment(
             this,
@@ -106,6 +95,30 @@ class MainActivity : AppCompatActivity() {
             }
 
         })
+
+    }
+
+    // this method only needed if user has order_id to use check payment status
+    @RequiresApi(Build.VERSION_CODES.M)
+    private fun checkStatus(order_id: String) {
+
+        shurjopay?.verifyPayment(this,
+            order_id,
+            object : PaymentResultListener {
+                override fun onSuccess(success: ShurjopaySuccess) {
+                    Log.d(TAG, "onSuccess: debugMessage = ${success.debugMessage}")
+                }
+
+                override fun onFailed(exception: ShurjopayException) {
+                    Log.d(TAG, "onFailed: debugMessage = ${exception.debugMessage}")
+                }
+
+                override fun onBackButtonListener(exception: ShurjopayException): Boolean {
+                    Log.d(TAG, "onBackButton: debugMessage = ${exception.debugMessage}")
+                    return true
+                }
+
+            })
 
     }
 }
