+++ android-app-plugin/app/src/main/java/com/shurjopay/android/android_app_example/utils/Constants.kt
@@ -7,9 +7,9 @@ class Constants {
     @Keep
     companion object {
         // shujopay configuration parameter
-        const val SP_USER = "sp_sandbox"
-        const val SP_PASS = "pyyk97hu&6u6"
-        const val SHURJOPAY_API = "https://sandbox.shurjopayment.com/api/"
+        const val SP_USER = "shurjopay"
+        const val SP_PASS = "pyyk97hu&6u625"
+        const val SHURJOPAY_API = "https://www.dev.engine.shurjopayment.com"
     }
 
 }
