+++ android-app-plugin/app/src/main/res/values/colors.xml
