+++ android-app-plugin/app/src/main/res/values/strings.xml
