+++ android-app-plugin/app/src/test/java/com/shurjopay/android/android_app_example/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.shurjopay.android.android_app_kotlin_plugin
+package com.shurjopay.android.android_app_example
 
 import org.junit.Test
