+++ android-app-plugin/gradle.properties
