+++ android-app-plugin/settings.gradle
